requirements.txt
streamlit==1.31.0
google-generativeai==0.3.2
supabase==2.3.0
storage3==0.7.0
python-dotenv==1.0.0
